Moved from package `error_reporting` to `errorReporting`
https://github.com/php-perfect/ddev-intellij-plugin/pull/60#discussion_r883499472
