+++ src/main/java/de/php_perfect/intellij/ddev/InitPluginActivity.java
@@ -4,7 +4,7 @@ import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.extensions.ExtensionPointName;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.startup.StartupActivity;
-import de.php_perfect.intellij.ddev.error_reporting.SentrySdkInitializer;
+import de.php_perfect.intellij.ddev.errorReporting.SentrySdkInitializer;
 import de.php_perfect.intellij.ddev.state.DdevStateManager;
 import org.jetbrains.annotations.NotNull;
