+++ src/main/java/de/php_perfect/intellij/ddev/errorReporting/SentryErrorReporter.java
@@ -1,4 +1,4 @@
-package de.php_perfect.intellij.ddev.error_reporting;
+package de.php_perfect.intellij.ddev.errorReporting;
 
 import com.intellij.diagnostic.AbstractMessage;
 import com.intellij.diagnostic.IdeaReportingEvent;
