+++ src/main/java/de/php_perfect/intellij/ddev/errorReporting/ThreadSync.java
@@ -1,4 +1,4 @@
-package de.php_perfect.intellij.ddev.error_reporting;
+package de.php_perfect.intellij.ddev.errorReporting;
 
 import java.util.concurrent.atomic.AtomicReference;
