+++ src/main/resources/META-INF/plugin.xml
@@ -43,7 +43,7 @@
     </extensionPoints>
 
     <extensions defaultExtensionNs="com.intellij">
-        <errorHandler implementation="de.php_perfect.intellij.ddev.error_reporting.SentryErrorReporter"/>
+        <errorHandler implementation="de.php_perfect.intellij.ddev.errorReporting.SentryErrorReporter"/>
     </extensions>
 
     <extensions defaultExtensionNs="de.php_perfect.intellij.ddev">
